move tests, change cost
