+++ modules/sql-engine/src/test/java/org/apache/ignite/internal/sql/engine/planner/HashJoinPlannerTest.java
@@ -18,13 +18,27 @@
 package org.apache.ignite.internal.sql.engine.planner;
 
 import static org.apache.ignite.internal.sql.engine.planner.CorrelatedSubqueryPlannerTest.createTestTable;
+import static org.apache.ignite.internal.sql.engine.planner.JoinColocationPlannerTest.complexTbl;
+import static org.apache.ignite.internal.sql.engine.planner.JoinColocationPlannerTest.simpleTable;
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.notNullValue;
+import static org.hamcrest.MatcherAssert.assertThat;
 
+import java.util.List;
 import java.util.stream.Stream;
 import org.apache.calcite.plan.RelOptPlanner.CannotPlanException;
+import org.apache.calcite.plan.RelOptUtil;
+import org.apache.calcite.rel.RelNode;
+import org.apache.calcite.rel.core.Join;
+import org.apache.ignite.internal.sql.engine.rel.AbstractIgniteJoin;
 import org.apache.ignite.internal.sql.engine.rel.IgniteHashJoin;
+import org.apache.ignite.internal.sql.engine.rel.IgniteRel;
+import org.apache.ignite.internal.sql.engine.rel.IgniteSort;
 import org.apache.ignite.internal.sql.engine.schema.IgniteSchema;
 import org.apache.ignite.internal.sql.engine.schema.IgniteTable;
 import org.apache.ignite.internal.testframework.IgniteTestUtils;
+import org.jetbrains.annotations.Nullable;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -34,6 +48,94 @@ public class HashJoinPlannerTest extends AbstractPlannerTest {
     private static final String[] disabledRules = {"NestedLoopJoinConverter", "CorrelatedNestedLoopJoin", "MergeJoinConverter"};
 
     private static final String[] joinTypes = {"LEFT", "RIGHT", "INNER", "FULL OUTER"};
+
+    /**
+     * Hash join need to preserve left collation.
+     */
+    @Test
+    public void hashJoinCheckLeftCollationsPropagation() throws Exception {
+        IgniteTable tbl1 = simpleTable("TEST_TBL", DEFAULT_TBL_SIZE);
+        IgniteTable tbl2 = complexTbl("TEST_TBL_CMPLX");
+
+        IgniteSchema schema = createSchema(tbl1, tbl2);
+
+        String sql = "select t1.ID, t2.ID1 "
+                + "from TEST_TBL_CMPLX t2 "
+                + "join TEST_TBL t1 on t1.id = t2.id1 "
+                + "order by t2.ID1 NULLS LAST, t2.ID2 NULLS LAST";
+
+        // Only hash join
+        RelNode phys = physicalPlan(sql, schema, "NestedLoopJoinConverter",
+                "CorrelatedNestedLoopJoin", "MergeJoinConverter", "JoinCommuteRule");
+
+        AbstractIgniteJoin join = findFirstNode(phys, byClass(AbstractIgniteJoin.class));
+        List<RelNode> joinNodes = findNodes(phys, byClass(AbstractIgniteJoin.class));
+        List<RelNode> sortNodes = findNodes(phys, byClass(IgniteSort.class));
+
+        String invalidPlanMsg = "Invalid plan:\n" + RelOptUtil.toString(phys);
+
+        assertThat(invalidPlanMsg, sortNodes.size(), equalTo(0));
+        assertThat(invalidPlanMsg, joinNodes.size(), equalTo(1));
+        assertThat(invalidPlanMsg, join, notNullValue());
+    }
+
+    /**
+     * Hash join erase right collation.
+     */
+    @Test
+    public void hashJoinCheckRightCollations() throws Exception {
+        IgniteTable tbl1 = simpleTable("TEST_TBL", DEFAULT_TBL_SIZE);
+        IgniteTable tbl2 = complexTbl("TEST_TBL_CMPLX");
+
+        IgniteSchema schema = createSchema(tbl1, tbl2);
+
+        String sql = "select t1.ID, t2.ID1 "
+                + "from TEST_TBL t1 "
+                + "join TEST_TBL_CMPLX t2 on t1.id = t2.id1 "
+                + "order by t2.ID1 NULLS LAST, t2.ID2 NULLS LAST";
+
+        // Only hash join
+        IgniteRel phys = physicalPlan(sql, schema, "NestedLoopJoinConverter",
+                "CorrelatedNestedLoopJoin", "MergeJoinConverter", "JoinCommuteRule");
+
+        AbstractIgniteJoin join = findFirstNode(phys, byClass(AbstractIgniteJoin.class));
+
+        String invalidPlanMsg = "Invalid plan:\n" + RelOptUtil.toString(phys);
+
+        assertThat(invalidPlanMsg, join, notNullValue());
+        assertThat(invalidPlanMsg, sortOnTopOfJoin(phys), notNullValue());
+    }
+
+    @Test
+    public void hashJoinWinsOnLeftSkewedInput() throws Exception {
+        IgniteTable thinTbl = simpleTable("THIN_TBL", 10);
+        IgniteTable thickTbl = simpleTable("THICK_TBL", 100_000);
+
+        IgniteSchema schema = createSchema(thinTbl, thickTbl);
+
+        String sql = "select t1.ID, t1.ID2, t2.ID, t2.ID2 "
+                + "from THICK_TBL t1 " // left
+                + "join THIN_TBL t2 on t1.ID2 = t2.ID2 "; // right
+
+        assertPlan(sql, schema, nodeOrAnyChild(isInstanceOf(IgniteHashJoin.class)), "JoinCommuteRule");
+
+        sql = "select t1.ID, t1.ID2, t2.ID, t2.ID2 "
+                + "from THIN_TBL t1 " // left
+                + "join THICK_TBL t2 on t1.id = t2.id "; // right
+
+        assertPlan(sql, schema, nodeOrAnyChild(isInstanceOf(IgniteHashJoin.class).negate()), "JoinCommuteRule");
+    }
+
+    private static @Nullable IgniteSort sortOnTopOfJoin(IgniteRel root) {
+        List<IgniteSort> sortNodes = findNodes(root, byClass(IgniteSort.class)
+                .and(node -> node.getInputs().size() == 1 && node.getInput(0) instanceof Join));
+
+        if (sortNodes.size() > 1) {
+            throw new AssertionError("Unexpected count of sort nodes: exp<=1, act=" + sortNodes.size());
+        }
+
+        return sortNodes.isEmpty() ? null : sortNodes.get(0);
+    }
 
     /** Check that only appropriate conditions are acceptable for hash join. */
     @ParameterizedTest()
